--- src/interfaces/message.interface.ts
@@ -1,9 +1,0 @@
-import { Date, Document, Types } from "mongoose"
-
-export interface IMessage extends Document {
-  readonly _id: Types.ObjectId
-  readonly chatId: Types.ObjectId
-  readonly sender: Types.ObjectId
-  readonly timestamp: Date
-  readonly messageContent: string
-}
